added --appium switch
