+++ .gitignore
@@ -176,3 +176,4 @@ results-unittests/
 /key/BigQueryCredentialLink
 /key/nrobo-statistics-214a0f5d2c1b.json
 /key/experiments/pytest/
+node_modules/
