+++ conftest.py
@@ -41,6 +41,7 @@ import os.path as path
 import nrobo.cli.detection as detect
 
 from nrobo.util.constants import CONST
+from nrobo.appium import AUTOMATION_NAMES, CAPABILITY
 
 
 def update_pytest_life_cycle_log(life_cycle_item: str, item_type: str = "fixture"):
@@ -128,15 +129,23 @@ def add_capabilities_from_file(options):
        and return updated options"""
     from nrobo.util.common import Common
     from nrobo import NROBO_PATHS, Environment, EnvKeys
-    if detect.production_machine() and not detect.developer_machine():
-        capabilities = Common.read_yaml(NROBO_PATHS.EXEC_DIR / NROBO_PATHS.CAPABILITY_YAML)
-    else:
-        capabilities = Common.read_yaml(
-            NROBO_PATHS.NROBO_DIR / NROBO_PATHS.NROBO / NROBO_PATHS.CAPABILITY_YAML)
+    capabilities = Common.read_yaml(NROBO_PATHS.EXEC_DIR / NROBO_PATHS.CAPABILITY_YAML)
+
     for k, v in capabilities.items():
         options.set_capability(k, v)
 
     return options
+
+
+def get_appium_capabilities_from_file():
+    """Read appium capabilities from appium_capability.yaml file
+
+       return appium_capabilities"""
+    from nrobo.util.common import Common
+    from nrobo import NROBO_PATHS, Environment, EnvKeys
+    capabilities = Common.read_yaml(NROBO_PATHS.EXEC_DIR / NROBO_PATHS.CAPABILITY_APPIUM_YAML)
+
+    return capabilities
 
 
 def pytest_addoption(parser):
@@ -149,6 +158,13 @@ def pytest_addoption(parser):
     update_pytest_life_cycle_log("pytest_addoption", "hook")
 
     group = parser.getgroup("nrobo header options")
+    # nRoBo appium options
+    group.addoption(
+        f"--{nCLI.APPIUM}", help=f"Tells nRoBo to trigger via appium client",
+        action="store_true", default=False
+    )
+
+    # nRoBo webdriver options
     group.addoption(
         f"--{nCLI.BROWSER}", help="""
     Target browser name. Default is chrome.
@@ -169,6 +185,7 @@ def pytest_addoption(parser):
                     help="Take full page screenshot", action="store_true", default=False)
 
     # ini option
+    parser.addini(f"--{nCLI.APPIUM}", type='bool', help=f"Tells nRoBo to trigger via appium client")
     parser.addini(f"{nCLI.APP}", type="string",
                   help="Name of your app project under test")
     parser.addini(f"{nCLI.REPORT_TITLE}", type="string",
@@ -263,7 +280,20 @@ def driver(request):
                        NREPORT.LOG_DIR_DRIVER + os.sep + \
                        test_method_name + NREPORT.LOG_EXTENTION
 
-    if browser == Browsers.CHROME:
+    if bool(os.environ[EnvKeys.APPIUM]):
+        """get appium driver with given capabilities"""
+        from appium import webdriver as _webdriver
+
+        capabilities = get_appium_capabilities_from_file()
+
+        if capabilities[CAPABILITY.AUTOMATION_NAME] == AUTOMATION_NAMES.UI_AUTOMATION2:
+            """Create uiautomator2 driver instance"""
+            from appium.options.android import UiAutomator2Options
+
+            options = UiAutomator2Options().load_capabilities(capabilities)
+            _driver = _webdriver.Remote(_grid_server_url, options=options)
+
+    elif browser == Browsers.CHROME:
         """if browser requested is chrome"""
 
         options = webdriver.ChromeOptions()
@@ -285,6 +315,7 @@ def driver(request):
                                        service=ChromeService(
                                            ChromeDriverManager().install(),
                                            log_output=_driver_log_path))
+
     elif browser == Browsers.CHROME_HEADLESS:
         """if browser requested is chrome"""
 
@@ -337,6 +368,7 @@ def driver(request):
             #                            service=ChromeService(
             #                                ChromeDriverManager().install(),
             #                                log_output=_driver_log_path))
+
     elif browser == Browsers.SAFARI:
         """if browser requested is safari"""
 
@@ -382,6 +414,7 @@ def driver(request):
             """Get instance of local firefox driver"""
             _service = webdriver.FirefoxService(log_output=_driver_log_path, service_args=['--log', 'debug'])
             _driver = webdriver.Firefox(options=options, service=_service)
+
     elif browser == Browsers.EDGE:
         """if browser requested is microsoft edge"""
 
@@ -401,6 +434,7 @@ def driver(request):
             """Get instance of local firefox driver"""
             _service = webdriver.EdgeService(log_output=_driver_log_path)
             _driver = webdriver.Edge(options=options, service=_service)
+
     elif browser == Browsers.IE:
         """if browser requested is microsoft internet explorer"""
 
@@ -429,6 +463,7 @@ def driver(request):
             """Get instance of local firefox driver"""
             _service = webdriver.IeService(log_output=_driver_log_path)
             _driver = webdriver.Ie(options=options)
+
     else:
         from nrobo.cli.tools import console
         console.rule(f"[{STYLE.HLRed}]DriverNotConfigured Error!")
@@ -578,6 +613,7 @@ def pytest_configure(config):
 
     os.environ[EnvKeys.TITLE] = str(config.getoption(f'--{nCLI.REPORT_TITLE}')).replace(CONST.UNDERSCORE, CONST.SPACE)
     os.environ[EnvKeys.APP] = str(config.getoption(f'--{nCLI.APP}')).replace(CONST.UNDERSCORE, CONST.SPACE)
+    os.environ[EnvKeys.APPIUM] = str(config.getoption(f'--{nCLI.APPIUM}'))
 
     # add custom markers
     config.addinivalue_line("markers", "sanity: marks as sanity test")
@@ -632,17 +668,17 @@ def pytest_runtest_setup(item):
         fixtureinfo = None
 
     update_pytest_life_cycle_log_with_value(f"Function properties:\n"
-                                                                                f"name={item.name}\n"
-                                                                                f"parent={item.parent}\n"
-                                                                                f"config={item.config}\n"
-                                                                                f"callspec={callspec}\n"
-                                                                                f"callobj={callobj}\n"
-                                                                                f"keywords={item.keywords}\n"
-                                                                                f"session={item.session}\n"
-                                                                                f"fixtureinfo={fixtureinfo}\n"
-                                                                                f"originalname={item.originalname}\n"
-                                                                                f"filepath={item.fspath}\n"
-                                                                                f"docstring={item.__doc__}")
+                                            f"name={item.name}\n"
+                                            f"parent={item.parent}\n"
+                                            f"config={item.config}\n"
+                                            f"callspec={callspec}\n"
+                                            f"callobj={callobj}\n"
+                                            f"keywords={item.keywords}\n"
+                                            f"session={item.session}\n"
+                                            f"fixtureinfo={fixtureinfo}\n"
+                                            f"originalname={item.originalname}\n"
+                                            f"filepath={item.fspath}\n"
+                                            f"docstring={item.__doc__}")
 
     pprint(item.__doc__)
     pprint(item.config)
