+++ nrobo/appium/__init__.py
@@ -1,0 +1,25 @@
+"""
+=====================CAUTION=======================
+DO NOT DELETE THIS FILE SINCE IT IS PART OF NROBO
+FRAMEWORK AND IT MAY CHANGE IN THE FUTURE UPGRADES
+OF NROBO FRAMEWORK. THUS, TO BE ABLE TO SAFELY UPGRADE
+TO LATEST NROBO VERSION, PLEASE DO NOT DELETE THIS
+FILE OR ALTER ITS LOCATION OR ALTER ITS CONTENT!!!
+===================================================
+
+appium module holds nRoBo settings needed for
+working with appium framework.
+
+@author: Panchdev Singh Chauhan
+@email: erpanchdev@gmail.com
+"""
+
+
+class CAPABILITY:
+    AUTOMATION_NAME = 'automationName'
+
+
+class AUTOMATION_NAMES:
+    """Possible appium capabilities for automationName"""
+
+    UI_AUTOMATION2 = 'uiautomator2'
