+++ nrobo/browserConfigs/appium_capability.yaml
@@ -1,0 +1,7 @@
+platformName: Android
+automationName: uiautomator2
+deviceName: Android
+appPackage: com.android.settings
+appActivity: .Settings
+language: en
+locale: US
