+++ nrobo/cli/cli_constants.py
@@ -42,6 +42,8 @@ class nCLI:
         Make sure that the same option is also removed from the
         nCLI.ARGS dictionary too!!!"""
 
+    NPM = "npm"
+    APPIUM = "appium"
     INSTALL = "install"
     APP = "app"
     URL = "url"
@@ -66,6 +68,8 @@ class nCLI:
     MARKER = "marker"
 
     ARGS = {
+        NPM: NPM,
+        APPIUM: APPIUM,
         INSTALL: INSTALL,
         APP: APP,
         URL: URL,
@@ -103,3 +107,5 @@ class PACKAGES:
 
     NROBO = "nrobo"
     CLI = "cli"
+
+    APPIUM = ['appium']
