+++ nrobo/cli/install/__init__.py
@@ -182,7 +182,7 @@ def transfer_files_to_host_project() -> None:
         print(f"Installation complete")
 
 
-def install_nrobo(requirements_file: Optional[str] = None) -> None:
+def install_nrobo(requirements_file: Optional[str] = None, install_only: bool = False) -> None:
     """This will install nrobo framework and its dependencies on host system in the current directory
     from where nrobo command was executed in the Production environment.
 
@@ -210,6 +210,11 @@ def install_nrobo(requirements_file: Optional[str] = None) -> None:
         else:
             print(f"Requirements are not installed successfully!")
             return
+
+    if install_only:
+        # No need to install or upgrade framework
+        # Just return after installing requirements
+        return
 
     if detect.production_machine():
         """Install or upgrading framework on Production environment"""
