--- tests/no_gui/__init__.py
@@ -1,2 +1,0 @@
-def test_mul():
-    assert 2*3 == 6
