--- tests/selenium_wrapper_nrobo_methods/test_browser_switches.py
@@ -1,11 +1,0 @@
-from nrobo.framework.pages import Page
-
-
-class TestBrowserSwitches:
-
-    def test_anti_bot_undetected_chrome(self, driver, logger):
-        page = Page(driver=driver, logger=logger)
-
-        page.get("https://nowsecure.nl")
-
-        page.wait_for_a_while(3)
