+++ tests/web/no_gui/__init__.py
@@ -1,0 +1,2 @@
+def test_mul():
+    assert 2*3 == 6
