+++ tests/web/selenium_wrapper_nrobo_methods/test_action_chain_methods.py
@@ -1,0 +1,26 @@
+from selenium.webdriver.common.by import By
+
+from nrobo.framework.pages import Page
+
+
+class TestActionChainMethods:
+
+    def test_action_chain_reference(self, driver, logger):
+        """Example of action chain reference"""
+
+        page = Page(driver, logger)
+        page.get("https://the-internet.herokuapp.com/")
+
+        lnkDragAndDrop = (By.CSS_SELECTOR, "[href='/drag_and_drop']")
+        page.click(*lnkDragAndDrop)
+
+        boxHeaderA = (By.XPATH, "//header[text()='A']")
+        boxHeaderB = (By.XPATH, "//header[text()='B']")
+
+        eleHeaderA = page.find_element(*boxHeaderA)
+        eleHeaderB = page.find_element(*boxHeaderB)
+        page.action_chain().drag_and_drop(eleHeaderA, eleHeaderB).perform()
+
+        page.wait_for_a_while(4)
+
+
